chore: switch to esm
BREAKING CHANGE: only named exports are used, deep imports/requires are not possible

+++ packages/ipfs-repo-migrations/migrations/migration-11/index.js
@@ -1,6 +1,5 @@
-'use strict'
 
-const { Key } = require('interface-datastore')
+import { Key } from 'interface-datastore/key'
 
 const MFS_ROOT_KEY = new Key('/local/filesroot')
 
@@ -14,9 +13,11 @@ async function storeMfsRootInDatastore (backends, onProgress = () => {}) {
   await backends.root.open()
   await backends.datastore.open()
 
-  const root = await backends.root.get(MFS_ROOT_KEY)
-  await backends.datastore.put(MFS_ROOT_KEY, root)
-  await backends.root.delete(MFS_ROOT_KEY)
+  if (await backends.root.has(MFS_ROOT_KEY)) {
+    const root = await backends.root.get(MFS_ROOT_KEY)
+    await backends.datastore.put(MFS_ROOT_KEY, root)
+    await backends.root.delete(MFS_ROOT_KEY)
+  }
 
   await backends.datastore.close()
   await backends.root.close()
@@ -34,9 +35,11 @@ async function storeMfsRootInRoot (backends, onProgress = () => {}) {
   await backends.root.open()
   await backends.datastore.open()
 
-  const root = await backends.datastore.get(MFS_ROOT_KEY)
-  await backends.root.put(MFS_ROOT_KEY, root)
-  await backends.datastore.delete(MFS_ROOT_KEY)
+  if (await backends.datastore.has(MFS_ROOT_KEY)) {
+    const root = await backends.datastore.get(MFS_ROOT_KEY)
+    await backends.root.put(MFS_ROOT_KEY, root)
+    await backends.datastore.delete(MFS_ROOT_KEY)
+  }
 
   await backends.datastore.close()
   await backends.root.close()
@@ -45,7 +48,7 @@ async function storeMfsRootInRoot (backends, onProgress = () => {}) {
 }
 
 /** @type {import('../../src/types').Migration} */
-module.exports = {
+export const migration = {
   version: 11,
   description: 'Store mfs root in the datastore',
   migrate: storeMfsRootInDatastore,
